requirements.txt
torch
numpy
torchvision
python-chess
requests
pygame
